Move everything to top level dir
- to dev the app requires building now or a fancy regex for running all files (with `go run *...`) but must avoid tests. (discuss)
