--- config/testdata/invalid_config_author.json
@@ -1,3 +1,0 @@
-{
-    "email": "john.smith@example.com"
-}
