--- config/testdata/invalid_config_email.json
@@ -1,3 +1,0 @@
-{
-    "author": "John Smith"
-}
