--- config/testdata/valid_config_extras.json
@@ -1,5 +1,0 @@
-{
-    "author": "John Smith",
-    "email": "john.smith@example.com",
-    "something": "this should be ignored"
-}
