+++ helpers.go
@@ -4,7 +4,6 @@ import (
 	"bufio"
 	"fmt"
 	"os"
-	// "strconv"
 )
 
 // Prompt gets text input from the user and returns it as a string.
@@ -12,9 +11,16 @@ func Prompt(question string) string {
 	reader := bufio.NewReader(os.Stdin)
 	fmt.Print(question)
 	message, _ := reader.ReadString('\n')
-
 	return message
 }
+
+// ByDate implements sort.Interface for []Log based
+// on the Date field
+type ByDate []Log
+
+func (d ByDate) Len() int           { return len(d) }
+func (d ByDate) Swap(i, j int)      { d[i], d[j] = d[j], d[i] }
+func (d ByDate) Less(i, j int) bool { return d[i].Date > d[j].Date }
 
 // CheckArgs errors if user fails to provider correct num/types of os.Args.
 func CheckArgs(typesNeeded []string, args []string, commandName string) {
@@ -27,8 +33,4 @@ func CheckArgs(typesNeeded []string, args []string, commandName string) {
 
 	// TODO: iterate over `typesNeeded` and check each arg matches each typeneeded
 
-}
-
-func typeof(v interface{}) string {
-	return fmt.Sprintf("%T", v)
 }
