--- src/.beaconrc
@@ -1,4 +1,0 @@
-{
-    "author": "John Smith",
-    "email": "john.smith@example.com"
-}
