--- src/beacon_log.json
@@ -1,1 +1,0 @@
-[{"Date":1502571602,"Email":"john.smith@example.com","Author":"John Smith","Message":"This is my third message"},{"Date":1502571589,"Email":"john.smith@example.com","Author":"John Smith","Message":"This is my second message"},{"Date":1502571526,"Email":"john.smith@example.com","Author":"John Smith","Message":"This is my message"},{"Date":1502570879,"Email":"john.smith@example.com","Author":"John Smith","Message":"Second\n"},{"Date":1502570876,"Email":"john.smith@example.com","Author":"John Smith","Message":"First\n"},{"Date":1502584086,"Email":"john.smith@example.com","Author":"John Smith","Message":"hi hows it going"}]
